feat(xo-lite/vm system): find more data for vmGeneralInfo
